[UNSTABLE][WIP] Need to finish moving files into correct dirs, coverage, and adding batch support to more resources. #squashThis
