+++ app/lib/log_service.rb
