+++ lib/tasks/log_service_runner.rb
